requirements.txt
pandas==1.4.2
mysql-connector-python==8.0.29
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pyYAML==6.0
spotipy==2.19.0
requests==2.27.1